Added Render Primitives and Other Goodies
Added interfaces for vertex arrays, vertex buffers, and index buffers,
as well as basic abstraction for shaders. Also added a the beginnings
of a renderer pipeline with the Renderer being very high level, dealing
with the scene, whereas the rendererAPI and render commands deal with
slightly more low level stuff.

Also added a typedef for unsigned ints, uint can now be used.