Added Basic Scene System
Scenes are going to be the base of every game in Oyl3D. The Application
will be responsible for managing which single scene is currently being
run, and the scene manages it's layers which in turn manages their
GameObjects. The scene may be able to manage game objects in the future,
but this is not explicitly planned.